feat(desktop): model mappings page adjust
