+++ desktop/src/locales/en.ts
@@ -103,7 +103,7 @@ const en: Locale = {
     logout: 'Log out',
     settings: 'Settings',
     appSettings: 'App settings',
-    advancedConfig: 'Advanced config',
+    advancedConfig: 'Model mappings',
     changeAuth: 'Auth config',
     restart: 'Restart',
     restarting: 'Restarting…',
@@ -158,11 +158,11 @@ const en: Locale = {
     saving: 'Saving…',
   },
   advancedConfig: {
-    title: 'Advanced config',
+    title: 'Model mappings',
     subtitle: 'Manage the shared model mappings through the running service API so new requests to /v1/messages, /v1/messages/count_tokens, /v1/responses, and /v1/chat/completions pick up changes immediately.',
     back: 'Back to dashboard',
     modelMappingsTitle: 'Model mappings',
-    modelMappingsDesc: 'Map the requested model name to another target model. The same mappings are shared across supported interfaces, and targets may be regular model IDs or provider/model aliases.',
+    modelMappingsDesc: 'Map a requested model name to a target model ID or provider/model alias. Shared across all interfaces.',
     scopeLabel: 'Scope',
     scopeNote: 'One shared mapping set applies to POST /v1/messages, POST /v1/messages/count_tokens, POST /v1/responses, and POST /v1/chat/completions. Matching is exact and the app will not trim or normalize the model name for you.',
     restartNote: 'Changes are applied to the running service immediately after a successful save.',
@@ -177,7 +177,7 @@ const en: Locale = {
     saved: 'Saved through the service API. New requests to /v1/messages, /v1/messages/count_tokens, /v1/responses, and /v1/chat/completions will use the updated mappings immediately.',
     loadFailed: 'Failed to load model mappings',
     saveFailed: 'Failed to save model mappings',
-    serverRequired: 'Start the local service before loading or saving advanced config through the API.',
+    serverRequired: 'Start the local service before loading or saving model mappings through the API.',
     validationIncomplete: 'Each mapping row must include both a requested model and a target model.',
     validationDuplicate: 'Duplicate requested model: {{model}}',
   },
