+++ desktop/src/locales/zh.ts
@@ -103,7 +103,7 @@ const zh: Locale = {
     logout: '注销',
     settings: '设置',
     appSettings: '应用设置',
-    advancedConfig: '高级配置',
+    advancedConfig: '模型映射',
     changeAuth: '授权配置',
     restart: '重启',
     restarting: '重启中…',
@@ -158,11 +158,11 @@ const zh: Locale = {
     saving: '保存中…',
   },
   advancedConfig: {
-    title: '高级配置',
+    title: '模型映射',
     subtitle: '通过正在运行的服务接口管理共享的模型映射，保存后新的 `POST /v1/messages`、`POST /v1/messages/count_tokens`、`POST /v1/responses` 和 `POST /v1/chat/completions` 请求都会立即使用最新规则。',
     back: '返回看板',
     modelMappingsTitle: '模型映射',
-    modelMappingsDesc: '把请求里的模型名映射成另一个目标模型。所有支持的接口共用同一份映射，目标值既可以是普通模型 ID，也可以是 `provider/model`。',
+    modelMappingsDesc: '将请求模型名映射为另一个模型 ID 或 provider/model。所有接口共用。',
     scopeLabel: '作用范围',
     scopeNote: '同一份映射会统一作用于 `POST /v1/messages`、`POST /v1/messages/count_tokens`、`POST /v1/responses` 和 `POST /v1/chat/completions`。匹配为精确匹配，应用不会自动帮你裁剪空格或规范化大小写。',
     restartNote: '保存成功后会立刻刷新当前服务进程中的配置缓存。',
@@ -177,7 +177,7 @@ const zh: Locale = {
     saved: '已通过服务接口保存。后续新的 `POST /v1/messages`、`POST /v1/messages/count_tokens`、`POST /v1/responses` 和 `POST /v1/chat/completions` 请求都会立即使用最新映射。',
     loadFailed: '加载模型映射失败',
     saveFailed: '保存模型映射失败',
-    serverRequired: '请先启动本地服务，再通过接口读取或保存高级配置。',
+    serverRequired: '请先启动本地服务，再通过接口读取或保存模型映射。',
     validationIncomplete: '每一行映射都必须同时填写请求模型和目标模型。',
     validationDuplicate: '请求模型重复：{{model}}',
   },
